block: Disable random entropy generation
According to https://wiki.samat.org/SSD - With rotating disks, the amount of time to write a block is used as source of entropy. With SSDs, the time is so fast that it no longer is a good entropy source. The code path to verify randomness and add this data to the entropy pool just serves as a slowdown